Refactor: Create stats file handling stats server
