+++ src/provider.ts
@@ -7,15 +7,16 @@ dotenv.config();
 
 let chainId: ChainID;
 
-export const setChainId = (newChainId: number): void => { chainId = newChainId; };
+export const setChainId = (newChainId: number): void => {
+  chainId = newChainId;
+};
 export const getChainId = (): ChainID => chainId;
 
 const provider = new providers.JsonRpcProvider(process.env.CHAIN_RPC_ENDPOINT);
 
-provider.getNetwork().then(
-  ({ chainId: currentChainId }) => {
-    setChainId(currentChainId);
-  },
-);
+export const initialiseProvider = async (): Promise<void> => {
+  const { chainId } = await provider.getNetwork();
+  setChainId(chainId);
+};
 
 export default provider;
